plus_open/close: mmap() top delta BAT
If we are going to write to the top delta, it's good to have its BAT
to be mmapped so we can effectively update it. Let's mmap() BAT in
plus_open() and munmap() it in plus_close().

While at it, mark the top delta as dirty when opened rw, and mark it
as clean at closing.

Signed-off-by: Kir Kolyshkin <kir@sacred.ru>